# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"AuthTokenResponse": {
				"properties": {
					"accessToken": {
						"type": "string"
					}
				},
				"required": [
					"accessToken"
				],
				"type": "object"
			},
			"UnauthorizedError": {
				"properties": {
					"name": {
						"type": "string",
						"enum": [
							"UNAUTHORIZED"
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"cause": {},
					"description": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"message",
					"cause",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthTokenRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Profile": {
				"properties": {
					"postalCode": {
						"type": "string",
						"nullable": true
					},
					"country": {
						"type": "string",
						"nullable": true
					},
					"state": {
						"type": "string",
						"nullable": true
					},
					"city": {
						"type": "string",
						"nullable": true
					},
					"website": {
						"type": "string",
						"nullable": true
					},
					"bio": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"postalCode",
					"country",
					"state",
					"city",
					"website",
					"bio",
					"name"
				],
				"type": "object",
				"description": "Model Profile"
			},
			"User": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"profile": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Profile"
							}
						],
						"nullable": true
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"profile",
					"password",
					"email",
					"id"
				],
				"type": "object",
				"description": "Model User"
			},
			"BaseResponse_User-Array_": {
				"properties": {
					"content": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"content"
				],
				"type": "object"
			},
			"GetAllResponse": {
				"$ref": "#/components/schemas/BaseResponse_User-Array_"
			},
			"BaseResponse_User_": {
				"properties": {
					"content": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"content"
				],
				"type": "object"
			},
			"GetResponse": {
				"$ref": "#/components/schemas/BaseResponse_User_"
			},
			"NotFoundError": {
				"properties": {
					"name": {
						"type": "string",
						"enum": [
							"NOT_FOUND"
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"cause": {},
					"description": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"message",
					"cause",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SearchRequest": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UnprocessableEntityError": {
				"properties": {
					"name": {
						"type": "string",
						"enum": [
							"UNPROCESSABLE_ENTITY"
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"cause": {},
					"description": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"message",
					"cause",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterUserRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChangePassword": {
				"properties": {
					"currentPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"currentPassword",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChangeEmailRequest": {
				"properties": {
					"newEmail": {
						"type": "string"
					}
				},
				"required": [
					"newEmail"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EditProfileRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "playground",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/token": {
			"post": {
				"operationId": "Authenticate",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthTokenResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthTokenRequest"
							}
						}
					}
				}
			}
		},
		"/users": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetAllResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/users/{entityId}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/search": {
			"post": {
				"operationId": "GetByEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SearchRequest"
							}
						}
					}
				}
			}
		},
		"/users/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created"
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnprocessableEntityError"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterUserRequest"
							}
						}
					}
				}
			}
		},
		"/users/{entityId}/change-password": {
			"patch": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": ""
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnprocessableEntityError"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnprocessableEntityError"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePassword"
							}
						}
					}
				}
			}
		},
		"/users/{entityId}/change-email": {
			"patch": {
				"operationId": "ChangeEmail",
				"responses": {
					"200": {
						"description": ""
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnprocessableEntityError"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnprocessableEntityError"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangeEmailRequest"
							}
						}
					}
				}
			}
		},
		"/users/{entityId}/profile": {
			"put": {
				"operationId": "EditProfile",
				"responses": {
					"200": {
						"description": ""
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnprocessableEntityError"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnprocessableEntityError"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EditProfileRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}